Close collision-query enemyDistance precision fork (P5 Commit 3)
The snapshot path (production, dodge-collision-world) stores
`enemyDistance` as Float32Array — values are quantised on write and
quantised again on Float32Array read-back. The fallback path (tests
without `createLocalSnapshot`) returned `environment.enemyClearance` at
full-double precision. Downstream `safety.softCost` and `terminalCost`
differed by O(1e-12) to ~2.3e-7 between the two paths. Below rankScore's
selection epsilon most of the time, but not always — degenerate ties
flip.

One-line fix in the fallback branch of `createCollisionQuery`:

    enemyDistance: (x, y) => Math.fround(
      environment.enemyClearance?.(x, y) ?? Infinity,
    )

Rationale for `Math.fround` over upgrading dodge-collision-world's
Float32Array to Float64Array:
* Minimal invasion — one line here vs 3+ sites in dodge-collision-world.
* No memory increase (~180 KB per snapshot avoided).
* Test paths now match production precision — closes the fork.
* Goal is deterministic replay, not higher precision. Both paths
  converge on Float32 semantics.

Test:
* `createCollisionQuery fallback quantises enemyDistance to Float32
  for snapshot parity` — builds an environment whose `enemyClearance`
  returns a full-double value that doesn't round-trip through Float32;
  asserts `result.minimumEnemyClearance` matches `Math.fround(same)`.
  Sanity assertion first confirms the test value actually differs
  under Math.fround.

Files touched:
* HeadlessClient/headless-client/src/dodge-trajectory-planner.ts
* HeadlessClient/headless-client/test/dodge-trajectory-planner.test.ts

Base: upstream/main @ 412ead9. Related spec:
docs/superpowers/specs/2026-07-19-replay-hardening-design.md

Executed-By: LO

+++ HeadlessClient/headless-client/src/dodge-trajectory-planner.ts
@@ -1752,7 +1752,13 @@ function createCollisionQuery(
       blocked: (x, y) => rawBlocked(x, y)
         && !(allowStartingTile && isStartingTile(x, y)),
       damagingFloor: () => 0,
-      enemyDistance: (x, y) => environment.enemyClearance?.(x, y) ?? Infinity,
+      // Snapshot path (production, dodge-collision-world) stores enemyDistance
+      // as Float32Array; the fallback (tests) returns full-double
+      // environment.enemyClearance. Precision fork through evaluateEdge's
+      // softCost + terminalCost could flip degenerate rankScore ties between
+      // the two paths. Quantise to Float32 here so the fallback matches
+      // snapshot semantics — closes the fork at negligible cost.
+      enemyDistance: (x, y) => Math.fround(environment.enemyClearance?.(x, y) ?? Infinity),
     };
   }
   const indexAt = (x: number, y: number): number => {
